+++ missiongen/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.8.5"
+__version__ = "1.8.6"
 
 from .recipe import Recipe
 from .builder import StarterBuilder, generate